requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
google-genai>=1.47.0
python-dotenv==1.0.1
pydantic==2.10.4
aiosqlite==0.20.0
httpx>=0.27.0
python-multipart>=0.0.9
Pillow>=11.0.0
pillow-heif>=1.1.0
